chore(wellinks-extension): renames sdk and graphqlClient for Wellinks
